Revert "refactor: change and improve api endpoint routes"
This reverts commit 0f9dd6f7d8c72950418a257a3508c8d0c3a98a53.
