+++ src/routes/artifact.task.routes.ts
@@ -1,13 +1,13 @@
 /* eslint-disable @typescript-eslint/no-misused-promises */
-import { ArtifactTaskController } from "@controllers/api/artifact.task.controller";
+import { ArtifactTaskController } from "@controllers/artifact.task.controller";
 import { container } from "tsyringe";
 import express from "express";
 
-export const taskApiRouter = express.Router();
+export const artifactTaskRouter = express.Router();
 const artifactTaskController = container.resolve(ArtifactTaskController);
 
-taskApiRouter.route("/").post(artifactTaskController.createTask.bind(artifactTaskController));
+artifactTaskRouter.route("/").post(artifactTaskController.createTask.bind(artifactTaskController));
 
-taskApiRouter
-  .route("/:taskId/status")
+artifactTaskRouter
+  .route("/status")
   .get(artifactTaskController.getTaskStatus.bind(artifactTaskController));
